Made counter for Bun-elements, fix conflict between rootReducer&Type Script
